general : seperated estimator
